pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["pylive"]


[project]
name = "pylive"
version = "0.1.0"
description = "Live coding expreiments"
authors = [
    {name = "Andras Zalavari", email = "zalavariandris@gmail.com.com"}
]
readme = "README.md"
keywords = ["python", "livecode", "livecoding"]
requires-python = ">=3.8"

dependencies = [
    "watchfiles ~= 0.24.0",
    "PySide6 ~= 6.7.0",
    "rope ~= 1.13.0",
    "pygments ~= 2.18.0",
    "IPython ~= 8.29.0",
    "numpy ~= 2.1.0",
    "humanize ~= 4.11.0",
    "NodeGraphQt@git+https://github.com/C3RV1/NodeGraphQt-PySide6",
    "PyOpenGL ~= 3.1",
    "moderngl == 5.12.0",
    "pyglm == 2.7.3",
    "trimesh == 4.5.2",
    "qtconsole == 5.6.1",
    "Panda3D == 1.10.15"
]
